fix: split files to resolve file service failed to load
