+++ packages/maskbook/src/plugins/PluginUI.ts
@@ -11,24 +11,27 @@ import { GitcoinPluginDefine } from './Gitcoin/define'
 import { RedPacketPluginDefine } from './RedPacket/define'
 import { PollsPluginDefine } from './Polls/define'
 import { StorybookPluginDefine } from './Storybook/define'
-import { FileServicePluginDefine } from './FileService/define'
+import { FileServicePluginDefine } from './FileService/UI-define'
 import { TraderPluginDefine } from './Trader/define'
 import { Election2020PluginDefine } from './Election2020/define'
 import { TransakPluginDefine } from './Transak/define'
 import { COTM_PluginDefine } from './COTM/define'
 import { ITO_PluginDefine } from './ITO/define'
 import { NFTPluginsDefine } from './NFT/define'
+import { sideEffect } from '../utils/side-effects'
 
-plugins.add(EthereumPluginDefine)
-plugins.add(WalletPluginDefine)
-plugins.add(RedPacketPluginDefine)
-plugins.add(FileServicePluginDefine)
-plugins.add(ITO_PluginDefine)
-plugins.add(NFTPluginsDefine)
-if (Flags.gitcoin_enabled) plugins.add(GitcoinPluginDefine)
-if (Flags.poll_enabled) plugins.add(PollsPluginDefine)
-if (Flags.trader_enabled) plugins.add(TraderPluginDefine)
-if (Flags.transak_enabled) plugins.add(TransakPluginDefine)
-if (Flags.election2020_enabled) plugins.add(Election2020PluginDefine)
-if (Flags.COTM_enabled) plugins.add(COTM_PluginDefine)
-if (process.env.STORYBOOK) plugins.add(StorybookPluginDefine)
+sideEffect.then(() => {
+    plugins.add(EthereumPluginDefine)
+    plugins.add(WalletPluginDefine)
+    plugins.add(RedPacketPluginDefine)
+    plugins.add(FileServicePluginDefine)
+    plugins.add(ITO_PluginDefine)
+    plugins.add(NFTPluginsDefine)
+    if (Flags.gitcoin_enabled) plugins.add(GitcoinPluginDefine)
+    if (Flags.poll_enabled) plugins.add(PollsPluginDefine)
+    if (Flags.trader_enabled) plugins.add(TraderPluginDefine)
+    if (Flags.transak_enabled) plugins.add(TransakPluginDefine)
+    if (Flags.election2020_enabled) plugins.add(Election2020PluginDefine)
+    if (Flags.COTM_enabled) plugins.add(COTM_PluginDefine)
+    if (process.env.STORYBOOK) plugins.add(StorybookPluginDefine)
+})
